Decompose DashboardContext monolith (1607 -> 383 lines)
Extract types, job persistence, and queue runners into focused modules:
- dashboardTypes.ts: All shared type definitions and interfaces
- jobPersistence.ts: Job serialization, normalization, and IDB read/write
- useConversionQueue.ts: Conversion queue state, memory management, and runner
- useTranscriptionQueue.ts: Transcription queue, upload, and audio extraction

DashboardContext.tsx is now a thin orchestrator that composes these hooks.
